+++ tests/unit/base/test_config.py
@@ -1,24 +1,17 @@
 import unittest
-import sys
-from unittest.mock import patch
+from pysqa.base.config import QueueAdapterWithConfig
 
 
 class TestConfig(unittest.TestCase):
     def test_bad_queue_type(self):
-        from pysqa.base.config import QueueAdapterWithConfig
-
         with self.assertRaises(ValueError):
             QueueAdapterWithConfig(config={"queue_type": "error", "queues": {}})
 
     def test_pydantic_validation_missing_queues(self):
-        from pysqa.base.config import QueueAdapterWithConfig
-
         with self.assertRaises(ValueError):
             QueueAdapterWithConfig(config={"queue_type": "SLURM"})
 
     def test_pydantic_validation_wrong_type(self):
-        from pysqa.base.config import QueueAdapterWithConfig
-
         with self.assertRaises(ValueError):
             QueueAdapterWithConfig(
                 config={
@@ -28,8 +21,6 @@ class TestConfig(unittest.TestCase):
             )
 
     def test_pydantic_validation_extra_fields(self):
-        from pysqa.base.config import QueueAdapterWithConfig
-
         config = {
             "queue_type": "SLURM",
             "queue_primary": "sq",
@@ -39,20 +30,3 @@ class TestConfig(unittest.TestCase):
         qa = QueueAdapterWithConfig(config=config)
         self.assertEqual(qa.config["queues"]["sq"]["extra_field"], "allowed")
         self.assertEqual(qa.config["extra_top_level"], "also_allowed")
-
-    def test_no_pydantic_validation_extra_fields(self):
-        with patch.dict('sys.modules', {'pydantic': None}):
-            if 'pysqa.base.models.validate_config' in sys.modules:
-                del sys.modules['pysqa.base.models.validate_config']
-
-            from pysqa.base.config import QueueAdapterWithConfig
-
-            config = {
-                "queue_type": "SLURM",
-                "queue_primary": "sq",
-                "queues": {"sq": {"extra_field": "allowed"}},
-                "extra_top_level": "also_allowed",
-            }
-            qa = QueueAdapterWithConfig(config=config)
-            self.assertEqual(qa.config["queues"]["sq"]["extra_field"], "allowed")
-            self.assertEqual(qa.config["extra_top_level"], "also_allowed")
